Monitor alive monitor added to monitor list
